+++ lysl/src/main/java/com/whu/lysl/web/controllers/h5/SystemController.java
@@ -1,4 +1,4 @@
-package com.whu.lysl.web.controllers;
+package com.whu.lysl.web.controllers.h5;
 
 
 import com.alibaba.fastjson.JSON;
@@ -18,11 +18,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.SessionAttributes;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 import java.util.List;
 
 /**
